backend/services/backend/requirements.txt
#Basic
arrow==0.15.5

# Django
django==3.0.8
django-cors-headers==3.4.0
django-filter==2.3.0
django-redis==4.11.0
django-extensions==3.0.8
django-anymail==8.1
# django-mailgun==0.9.1

# Database
psycopg2-binary==2.8.5

# REST API
djangorestframework==3.11.0
djangorestframework-simplejwt==4.6.0

# Docs
drf-yasg==1.17.1

# ImageField support
Pillow==7.2.0

# Linters
pylint==2.5.0
pylint-django==2.0.15
vulture==2.1

# Tests
pytest==6.0.0
pytest-cov==2.8.1
pytest-django==3.9.0
pytest-factoryboy==2.0.3
pytest-lazy-fixture==0.6.3
factory-boy==2.12.0

# WSGI-server
gevent==20.9.0
gunicorn==20.0.4

# Static files
whitenoise==5.2.0
django-storages==1.11.1

django-extensions-shell==1.7.4.1
